refactor(iceberg): refactor file org
